RegExp for parsing and switches passed to `run` #2
In order to clean up code and be more aware of spaces, the command (ie:
non-switches parameters) are parsed with a RegExp so they understand paths even
when they contain spaces (with the help of double quotes). Also switches are
converted into an array (`child_process.spawn` compatible) and passed to the
`run` function in each action. This avoid an useless stringify/parse of the
switches.