Fix extracting to UNC paths
Node-7z extractFull was failing to extract to UNC paths (e.g. //somemachine/somepath)

The path.normalize call was taking in a string such as "-o//somemachine/somepath" and converting it to "-o/somemachine/somepath"
Stripping off the -o prior to normalisation and then adding it back in fixes the problem.