Testing prepare to publish, mac binaries will be pre included, others will be download at npm install time.
Mac OS 10.6 to 10.12 binaries will be pre download from rudix.org and included when this package is publish.

This package is begin prepared on windows platform and currently able to decode Mac OS *.pkg to get the binary files using 7zip and unar.

The windows and linux binary versions will be included at install time base on your host platform.