chore: Use streams/events as base mecanism
Promises are removed as the base mecanism of handling data from 7z cli
in favor of streams. It avoid dealing with bigs objects (eg: Resolve
with full list file name for big achive). And it's more the node-way.

ES6+ features are being supported

BREAKING CHANGES: Makes the codebase useless for the time beeing