chrore: Delete binaries
The point of this lib is to be a wrapper around 7zip cli. As well as I
can see how it would be convhenient for some users to have binaries
bundled with this lib, I don't think it is necessary.

Including binaries would have some undesired side effects:
 - Keeping up to date binaries on all platform
 - Huge augmentation in size
 - Lots of users having to keep thoose binaries even if they don't use
   them
 - Voiding the unix philosophy (do one thing well)
 - ...

To put it simple: It's the job of package managment systems like brew or
chocolatey, not node-7z